requirements.txt
Flask==2.2.2
authlib==1.1.0
python-dotenv==1.0.0
stytch==4.5.0
